requirements.txt
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.4.0
django-filter==23.3
drf-yasg==1.21.7